requirements.txt
alembic==1.11.1
amqp==2.6.1
aniso8601==9.0.1
arrow==0.13.1
asgiref==3.5.2
async-generator @ file:///home/ktietz/src/ci/async_generator_1611927993394/work
attrs==22.2.0
Automat==22.10.0
bcrypt==4.0.1
bidict==0.22.1
billiard==3.6.4.0
blinker @ file:///home/conda/feedstock_root/build_artifacts/blinker_1681349778161/work
boltons @ file:///C:/b/abs_707eo7c09t/croot/boltons_1677628723117/work
brotlipy==0.7.0
cachetools==5.3.0
celery==4.3.0
certifi @ file:///C:/b/abs_4a0polqwty/croot/certifi_1683875377622/work/certifi
cffi @ file:///C:/b/abs_49n3v2hyhr/croot/cffi_1670423218144/work
chardet==3.0.4
charset-normalizer @ file:///home/conda/feedstock_root/build_artifacts/charset-normalizer_1678108872112/work
click==7.1.2
colorama @ file:///C:/b/abs_a9ozq0l032/croot/colorama_1672387194846/work
conda==23.3.1
conda-package-handling @ file:///C:/b/abs_fcga8w0uem/croot/conda-package-handling_1672865024290/work
conda_package_streaming @ file:///C:/b/abs_0e5n5hdal3/croot/conda-package-streaming_1670508162902/work
constantly==15.1.0
cpe==1.2.1
cryptography @ file:///C:/b/abs_8ecplyc3n2/croot/cryptography_1677533105000/work
cssmin==0.2.0
cssselect==1.2.0
deepdiff==4.0.7
Deprecated==1.2.13
distlib==0.3.6
Django==4.1.4
dnspython==2.3.0
docopt==0.6.2
docx==0.2.4
docx2pdf==0.1.8
email-validator==1.1.0
exceptiongroup @ file:///home/conda/feedstock_root/build_artifacts/exceptiongroup_1678703645500/work
filelock==3.12.2
Flask==1.0.2
Flask-Admin==1.5.5
Flask-Assets==0.12
Flask-DebugToolbar==0.10.1
Flask-Gravatar==0.5.0
Flask-Limiter==1.4
Flask-Login==0.4.1
Flask-Mail==0.9.1
Flask-Migrate==3.0.1
flask-paginate==0.5.2
Flask-RESTful==0.3.8
Flask-SQLAlchemy==2.3.2
Flask-User==1.0.2.2
Flask-WTF==0.14.3
google-api-core==2.11.0
google-auth==2.17.3
google-cloud-vision==3.4.1
googleapis-common-protos==1.59.0
grpcio==1.54.0
grpcio-status==1.54.0
gunicorn==20.0.4
h11 @ file:///home/conda/feedstock_root/build_artifacts/h11_1664132893548/work
hijri-converter==2.3.1
hijridate==2.3.0
hyperlink==21.0.0
idna==2.10
importlib-metadata==4.13.0
importlib-resources==5.12.0
incremental==22.10.0
itemadapter==0.7.0
itemloaders==1.0.6
itsdangerous==1.1.0
Jinja2==2.11.3
jmespath==1.0.1
jsmin==2.2.2
jsonpatch @ file:///tmp/build/80754af9/jsonpatch_1615747632069/work
jsonpickle==3.0.1
jsonpointer==2.1
kombu==4.6.11
limits==3.5.0
lxml==4.9.2
Mako==1.2.4
MarkupSafe==1.1.1
menuinst @ file:///C:/ci/menuinst_1631733438520/work
nested-lookup==0.2.18
oauthlib @ file:///C:/b/abs_44riqpe5gg/croot/oauthlib_1679489644705/work
opencve==1.4.0
ordered-set==4.1.0
outcome @ file:///home/conda/feedstock_root/build_artifacts/outcome_1655217710638/work
packaging==22.0
parsel==1.7.0
passlib==1.7.4
pdf2image==1.16.3
Pillow==9.3.0
pipreqs==0.4.11
platformdirs==3.8.1
pluggy @ file:///C:/ci/pluggy_1648024580010/work
Protego==0.2.1
proto-plus==1.22.2
protobuf==4.23.0
psycopg2==2.9.6
psycopg2-binary==2.8.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycosat @ file:///C:/b/abs_4b1rrw8pn9/croot/pycosat_1666807711599/work
pycparser @ file:///tmp/build/80754af9/pycparser_1636541352034/work
PyDispatcher==2.0.6
PyJWT @ file:///home/conda/feedstock_root/build_artifacts/pyjwt_1683676063469/work
PyMuPDF==1.22.5
pyOpenSSL @ file:///home/conda/feedstock_root/build_artifacts/pyopenssl_1680037383858/work
PyPDF2==3.0.1
PySocks @ file:///C:/ci/pysocks_1605307512533/work
pyTelegramBotAPI==4.10.0
pytesseract==0.3.10
python-dateutil==2.8.2
python-docx==0.8.11
python-dotenv==1.0.0
python-engineio==4.4.1
python-socketio==5.8.0
pytz==2023.3
pywin32==305.1
queuelib==1.6.2
redis==3.3.11
requests==2.23.0
requests-file==1.5.1
requests-oauthlib @ file:///home/conda/feedstock_root/build_artifacts/requests-oauthlib_1643557462909/work
rsa==4.9
ruamel.yaml @ file:///C:/b/abs_30ee5qbthd/croot/ruamel.yaml_1666304562000/work
ruamel.yaml.clib @ file:///C:/b/abs_aarblxbilo/croot/ruamel.yaml.clib_1666302270884/work
Scrapy==2.7.1
selenium @ file:///home/conda/feedstock_root/build_artifacts/selenium_1682107665443/work
service-identity==21.1.0
six==1.16.0
sniffio @ file:///home/conda/feedstock_root/build_artifacts/sniffio_1662051266223/work
sortedcontainers @ file:///tmp/build/80754af9/sortedcontainers_1623949099177/work
SQLAlchemy==1.3.1
SQLAlchemy-Utils==0.33.11
sqlparse==0.4.3
telebot==0.0.5
tldextract==3.4.0
toolz @ file:///C:/b/abs_cfvk6rc40d/croot/toolz_1667464080130/work
tqdm @ file:///C:/b/abs_f76j9hg7pv/croot/tqdm_1679561871187/work
trio @ file:///C:/b/abs_fdt93vesd0/croot/trio_1677879581627/work
trio-websocket @ file:///home/conda/feedstock_root/build_artifacts/trio-websocket_1679232064156/work/dist
tweepy @ file:///home/conda/feedstock_root/build_artifacts/tweepy_1678386880489/work
Twisted==22.10.0
twisted-iocpsupport==1.0.2
typing_extensions==4.4.0
tzdata==2022.7
untangle==1.1.1
urllib3==1.25.11
vine==1.3.0
virtualenv==20.23.1
w3lib==2.1.1
waitress==2.1.2
Wand==0.6.11
webassets==2.0
webdriver-manager==3.8.6
Werkzeug==1.0.1
win-inet-pton @ file:///C:/ci/win_inet_pton_1605306162074/work
wincertstore==0.2
wrapt==1.15.0
wsproto @ file:///home/conda/feedstock_root/build_artifacts/wsproto_1661356345548/work
WTForms==3.0.1
yarg==0.1.9
zipp==3.15.0
zope.interface==5.5.2
zstandard==0.19.0
